+++ src/main/java/com/github/kazuhito_m/mysample/infrastructure/multidbsetting/ConfigForConnectToLogDatabase.java
@@ -1,4 +1,4 @@
-package com.github.kazuhito_m.mysample.infrastructure.datasource;
+package com.github.kazuhito_m.mysample.infrastructure.multidbsetting;
 
 import org.seasar.doma.boot.autoconfigure.DomaProperties;
 import org.seasar.doma.jdbc.Config;
@@ -6,23 +6,21 @@ import org.seasar.doma.jdbc.Naming;
 import org.seasar.doma.jdbc.dialect.Dialect;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Primary;
 import org.springframework.jdbc.datasource.TransactionAwareDataSourceProxy;
 import org.springframework.stereotype.Component;
 
 import javax.sql.DataSource;
 
-@Component("configForConnectToMainDatabase")
-@Primary
-public class ConfigForConnectToMainDatabase implements Config {
+@Component
+public class ConfigForConnectToLogDatabase implements Config {
     final DataSource dataSource;
     final Dialect dialect;
     final Naming naming;
 
-    public ConfigForConnectToMainDatabase(@Qualifier("dataSource") DataSource originalMainDataSource,
-                                          @Value("${doma.dialect}") String dialectName,
-                                          @Value("${doma.naming}") String namingRuleName) {
-        this.dataSource = new TransactionAwareDataSourceProxy(originalMainDataSource);
+    public ConfigForConnectToLogDatabase(@Qualifier("logDataSource") DataSource originalLogDataSource,
+                                         @Value("${doma.dialect}") String dialectName,
+                                         @Value("${doma.naming}") String namingRuleName) {
+        this.dataSource = new TransactionAwareDataSourceProxy(originalLogDataSource);
         this.dialect = DomaProperties.DialectType.valueOf(dialectName).create();
         this.naming = DomaProperties.NamingType.valueOf(namingRuleName).naming();
     }
