+++ src/main/java/com/github/kazuhito_m/mysample/infrastructure/multidbsetting/ConfigForConnectToMainDatabase.java
@@ -1,4 +1,4 @@
-package com.github.kazuhito_m.mysample.infrastructure.datasource;
+package com.github.kazuhito_m.mysample.infrastructure.multidbsetting;
 
 import org.seasar.doma.boot.autoconfigure.DomaProperties;
 import org.seasar.doma.jdbc.Config;
@@ -6,21 +6,23 @@ import org.seasar.doma.jdbc.Naming;
 import org.seasar.doma.jdbc.dialect.Dialect;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Primary;
 import org.springframework.jdbc.datasource.TransactionAwareDataSourceProxy;
 import org.springframework.stereotype.Component;
 
 import javax.sql.DataSource;
 
 @Component
-public class ConfigForConnectToLogDatabase implements Config {
+@Primary
+public class ConfigForConnectToMainDatabase implements Config {
     final DataSource dataSource;
     final Dialect dialect;
     final Naming naming;
 
-    public ConfigForConnectToLogDatabase(@Qualifier("logDataSource") DataSource originalLogDataSource,
-                                         @Value("${doma.dialect}") String dialectName,
-                                         @Value("${doma.naming}") String namingRuleName) {
-        this.dataSource = new TransactionAwareDataSourceProxy(originalLogDataSource);
+    public ConfigForConnectToMainDatabase(@Qualifier("dataSource") DataSource originalMainDataSource,
+                                          @Value("${doma.dialect}") String dialectName,
+                                          @Value("${doma.naming}") String namingRuleName) {
+        this.dataSource = new TransactionAwareDataSourceProxy(originalMainDataSource);
         this.dialect = DomaProperties.DialectType.valueOf(dialectName).create();
         this.naming = DomaProperties.NamingType.valueOf(namingRuleName).naming();
     }
